requirements.txt
ultralytics>=8.0.0
mediapipe>=0.10.0
opencv-python>=4.5.5.64
numpy>=1.22.0
Pillow>=9.0.0